Feature/move tables to main schema

+++ src/main/java/wmi/amu/edu/pl/pri/dto/ChapterVersionDto.java
@@ -16,5 +16,5 @@ public class ChapterVersionDto {
     private Date uploadTime;
     private String link;
     private String fileName;
-    private Integer uploaderId;
+    private Long uploaderId;
 }
