+++ src/main/java/wmi/amu/edu/pl/pri/repositories/SupervisorRepo.java
@@ -1,7 +1,7 @@
 package wmi.amu.edu.pl.pri.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wmi.amu.edu.pl.pri.models.SupervisorModel;
+import wmi.amu.edu.pl.pri.models.pri.SupervisorModel;
 
-public interface SupervisorRepo extends JpaRepository<SupervisorModel, Integer> {
+public interface SupervisorRepo extends JpaRepository<SupervisorModel, Long > {
 }
