+++ src/main/java/wmi/amu/edu/pl/pri/services/SupervisorService.java
@@ -3,9 +3,7 @@ package wmi.amu.edu.pl.pri.services;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import wmi.amu.edu.pl.pri.models.StudentModel;
-import wmi.amu.edu.pl.pri.models.SupervisorModel;
-import wmi.amu.edu.pl.pri.repositories.StudentRepo;
+import wmi.amu.edu.pl.pri.models.pri.SupervisorModel;
 import wmi.amu.edu.pl.pri.repositories.SupervisorRepo;
 
 import java.util.List;
@@ -21,7 +19,7 @@ public class SupervisorService {
     public List<SupervisorModel> getAllSupervisors(){
         return repo.findAll();
     }
-    public SupervisorModel getSupervisorById(Integer id){
+    public SupervisorModel getSupervisorById(Long id){
         Optional<SupervisorModel> optionalFile = repo.findById(id);
         if(optionalFile.isPresent()){
             return optionalFile.get();
